% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{amplitude_phase_adjust}
\alias{amplitude_phase_adjust}
\title{amplitude_phase_adjust}
\usage{
amplitude_phase_adjust(coefs, n, intercept)
}
\arguments{
\item{coefs}{matrix of data}

\item{n}{number of values per trace}

\item{intercept}{whether to include an intercept}
}
\value{
phase and amplitude
}
\description{
adjust curves based on regression stats
}
