% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{amplitude_phase_adjust2}
\alias{amplitude_phase_adjust2}
\title{convert_to_amplitude_phase}
\usage{
amplitude_phase_adjust2(coefs, n, intercept)
}
\arguments{
\item{coefs}{matrix of data}

\item{n}{number of values per trace}

\item{intercept}{whether to include an intercept}
}
\value{
phase and amplitude
}
\description{
fit sin curves to atv data.
}
