% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scale.R
\name{color_scale}
\alias{color_scale}
\title{color_scale}
\usage{
color_scale(n = 100)
}
\arguments{
\item{n}{number of colors}
}
\value{
set of colors (dark blue, dark red, yellow like wellCAD)
}
\description{
color_scale
}
\examples{
color_scale(10)
}
