% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert_to_amplitude_phase}
\alias{convert_to_amplitude_phase}
\title{convert_to_amplitude_phase}
\usage{
convert_to_amplitude_phase(coefs)
}
\arguments{
\item{coefficient}{matrix of data}
}
\value{
phase and amplitude
}
\description{
fit sin curves to atv data.
}
