% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raster.R
\name{create_raster}
\alias{create_raster}
\title{------------------------------------------------------------------------------
create_raster}
\usage{
create_raster(x, elev = FALSE)
}
\arguments{
\item{x}{the matrix from a .las file}

\item{elev}{are values in elevation}
}
\value{
raster
}
\description{
------------------------------------------------------------------------------
create_raster
}
