% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_sin_cos}
\alias{create_sin_cos}
\title{create_sin_cos}
\usage{
create_sin_cos(n_rows, n_curves, intercept = TRUE)
}
\arguments{
\item{n_rows}{the number of rows in the atv image}

\item{n_curves}{the number of sin/cos curves to use}
}
\value{
fitting coefficients
}
\description{
fit sin curves to atv data
}
