% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangular_filter.R
\name{rectangular_filter}
\alias{rectangular_filter}
\title{rectangular_filter}
\usage{
rectangular_filter(x, dif = 4, k_h = 20, k_v = 4, ...)
}
\arguments{
\item{x}{matrix of data}

\item{dif}{difference between values and the moving quantile}

\item{k_h}{width of horizontal moving window;}

\item{k_v}{width of vertical moving window;}

\item{...}{arguments to pass to runquantile}
}
\value{
matrix of values masked with NA
}
\description{
rectangular_filter
}
