% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_filter.R
\name{row_filter}
\alias{row_filter}
\title{row_filter}
\usage{
row_filter(x, dif = 4, k = 21, probs = 0.5, ...)
}
\arguments{
\item{x}{matrix of data}

\item{dif}{difference between values and the moving quantile}

\item{k}{width of moving window; must be an integer between one and n}

\item{probs}{numeric vector of probabilities with values in [0,1] range used by runquantile. For example Probs=c(0,0.5,1) would be equivalent to running runmin, runmed and runmax. Same as probs in quantile function.}

\item{...}{arguments to pass to runquantile}
}
\description{
row_filter
}
