% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_signal.R
\name{shift_signal}
\alias{shift_signal}
\title{shift_signal}
\usage{
shift_signal(x, phase, direction = 1)
}
\arguments{
\item{x}{matrix}

\item{phase}{the amount to shift}

\item{direction}{the direction to shift}
}
\value{
shifted signal
}
\description{
shift_signal
}
