% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sin_fit_ols_parallel}
\alias{sin_fit_ols_parallel}
\title{sin_fit_ols_parallel}
\usage{
sin_fit_ols_parallel(y, n_curves)
}
\arguments{
\item{y}{atv matrix of data}

\item{n_curves}{the number of sin/cos curves to use}
}
\value{
fitting coefficients
}
\description{
fit sin curves to atv data.
}
