% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_depth.R
\name{subset_depth}
\alias{subset_depth}
\title{subset_depth}
\usage{
subset_depth(x, top, bot, elev = FALSE)
}
\arguments{
\item{x}{raster}

\item{top}{top value}

\item{bot}{bottom value}

\item{elev}{whether values are in elevation or depth}
}
\value{
subsetted raster
}
\description{
subset_depth
}
