% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radian_degree_conversion.R
\name{to_degree}
\alias{to_degree}
\title{to_degree}
\usage{
to_degree(x)
}
\arguments{
\item{x}{angle in radians}
}
\value{
angle in degrees
}
\description{
to_degree
}
\examples{
to_degree(2)
}
