% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radian_degree_conversion.R
\name{to_radian}
\alias{to_radian}
\title{to_radian}
\usage{
to_radian(x)
}
\arguments{
\item{x}{angle in degrees}
}
\value{
angle in radians
}
\description{
to_radian
}
\examples{
to_radian(180)

}
